solved some intellij warnings

+++ src/main/java/sfa/classification/BOSSEnsembleClassifier.java
@@ -63,7 +63,7 @@ public class BOSSEnsembleClassifier extends Classifier {
 
     if (DEBUG) {
       System.out.println(score.toString());
-      outputResult((int) score.training, startTime, testSamples.length);
+      outputResult(score.training, startTime, testSamples.length);
       System.out.println("");
     }
